Add initial finalization step
