+++ package.json
@@ -45,7 +45,7 @@
   "dependencies": {
     "@babel/code-frame": "7.12.11",
     "@eslint/eslintrc": "^0.4.2",
-    "@humanwhocodes/config-array": "^0.3.2",
+    "@humanwhocodes/config-array": "^0.4.0",
     "ajv": "^6.10.0",
     "chalk": "^4.0.0",
     "cross-spawn": "^7.0.2",
